@@ -1,14 +1,10 @@
-import BootstrapVue from 'bootstrap-vue';
-import Vue from 'vue';
-import App from './App.vue';
-import router from './router';
-import 'bootstrap/dist/css/bootstrap.css';
+import { createApp } from 'vue'
+import App from './App.vue'
+import router from './router'
+import 'bootstrap/dist/css/bootstrap.css'
 
-Vue.use(BootstrapVue);
+const app = createApp(App)
 
-Vue.config.productionTip = false;
+app.use(router)
 
-new Vue({
-  router,
-  render: (h) => h(App),
-}).$mount('#app');
+app.mount('#app')
