@@ -1,8 +1,10 @@
 <template>
-  <div id="app">
-    <router-view/>
-  </div>
+  <RouterView />
 </template>
+
+<script setup>
+import { RouterView } from 'vue-router'
+</script>
 
 <style>
 #app {
