@@ -1,15 +1,11 @@
-import Vue from 'vue';
-import Router from 'vue-router';
-import Books from '../components/Books.vue';
-import OrderCanceled from '../components/OrderCanceled.vue';
-import OrderSuccess from '../components/OrderSuccess.vue';
-import Ping from '../components/Ping.vue';
+import { createRouter, createWebHistory } from 'vue-router'
+import Books from '../components/Books.vue'
+import OrderCanceled from '../components/OrderCanceled.vue'
+import OrderSuccess from '../components/OrderSuccess.vue'
+import Ping from '../components/Ping.vue'
 
-Vue.use(Router);
-
-export default new Router({
-  mode: 'history',
-  base: process.env.BASE_URL,
+const router = createRouter({
+  history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     {
       path: '/',
@@ -18,8 +14,8 @@ export default new Router({
     },
     {
       path: '/ping',
-      name: 'Ping',
-      component: Ping,
+      name: 'ping',
+      component: Ping
     },
     {
       path: '/success',
@@ -31,5 +27,7 @@ export default new Router({
       name: 'OrderCanceled',
       component: OrderCanceled,
     },
-  ],
-});
+  ]
+})
+
+export default router
